Minor changes in indexer
